Bug fixes
There was a bug in admin_sign_ups where the variable keeping track of
the number of soldiers assigned wasn't being incremented, so only the
last assigned soldier was being added.

class.phpbb_interaction, group_remove_users function had a small bug
where the $auth var was not declared as a global var at the top of the
function.